add Text docs
