+++ packages/web-ui/src/Text/Text.docs.mdx
@@ -1,12 +1,12 @@
 import { Meta, Canvas, Story } from '@storybook/blocks';
-import Heading from './Heading.tsx';
-import * as Stories from './Heading.stories';
+import Text from './Text.tsx';
+import * as Stories from './Text.stories';
 
 <Meta of={Stories} />
 
-# Heading
+# Text
 
-Heading renders the primary UW font, to be used for heading-level typography.
+Heading renders the secondary UW font, Work Sans, to be used for body copy.
 
 <Canvas>
   <Story name="Workshop" />
@@ -14,7 +14,7 @@ Heading renders the primary UW font, to be used for heading-level typography.
 
 ## Alternatives
 
-- [Text](?path=/docs/components-text--documentation) - For body text
+- [Heading](?path=/docs/components-heading--documentation) - For heading-level text
 - [Typography]() - For custom typography
 
 ## Semantic HTML
