+++ packages/web-ui/src/Text/index.ts
@@ -1,0 +1,2 @@
+export { default, textVariantMapping } from './Text';
+export type { DefaultTextComponent, CustomTextProps, TextTypeMap, TextProps } from './Text';
