+++ packages/web-ui/src/Typography/Typography.tsx
@@ -1,25 +1,17 @@
 import { forwardRef } from 'react';
 import MuiTypography from '@mui/material/Typography';
 import { OverridableComponent } from '@mui/material/OverridableComponent';
-import { dataAttributes } from '../utils';
 import type { TypographyProps as MuiTypographyProps } from '@mui/material/Typography';
 import type { OverrideProps } from '@mui/material/OverridableComponent';
 import Heading, { HeadingProps, headingVariantMapping } from '../Heading';
+import Text, { TextProps, textVariantMapping } from '../Text';
 
 type DefaultTypographyComponent = 'p';
-type DefaultTextComponent = 'span';
-
-export const textVariantMapping: Record<string, string> = {
-  subtitle: 'p',
-  body: 'p',
-  legalNote: 'p',
-  caption: 'span',
-};
 
 interface CustomTypographyProps {
   color?: 'primary' | 'secondary' | 'success' | 'error';
-  bold?: boolean;
   variant: MuiTypographyProps['variant'];
+  component?: React.ElementType;
 }
 
 interface TypographyTypeMap<D extends React.ElementType = DefaultTypographyComponent, P = {}> {
@@ -32,50 +24,17 @@ type TypographyProps<
   P = {}
 > = OverrideProps<TypographyTypeMap<D, P>, D>;
 
-interface CustomTextProps {
-  /**
-   * Applies the theme typography styles.
-   */
-  variant: 'subtitle' | 'body' | 'legalNote' | 'caption';
-  color?: 'primary' | 'success' | 'error';
-}
-
-interface TextTypeMap<D extends React.ElementType = DefaultTextComponent, P = {}> {
-  props: Omit<TypographyProps<D, P>, 'variant' | 'fontWeight' | 'color'> & CustomTextProps;
-  defaultComponent: D;
-}
-
-type TextProps<D extends React.ElementType = DefaultTextComponent, P = {}> = OverrideProps<
-  TextTypeMap<D, P>,
-  D
->;
-
-const Text = forwardRef(function Text({ color = 'primary', bold = false, ...props }, ref) {
-  const dataAttributeProps = {
-    [`data-${dataAttributes[color]}`]: true,
-    [`data-${dataAttributes.bold}`]: bold,
-  };
-  return (
-    <MuiTypography
-      ref={ref}
-      variantMapping={textVariantMapping}
-      {...props}
-      {...dataAttributeProps}
-    />
-  );
-}) as OverridableComponent<TextTypeMap>;
-
 const Typography = forwardRef(function Typography(
-  { color = 'primary', bold = false, variant, ...props },
+  { color = 'primary', variant, component = 'p', ...props },
   ref
 ) {
   if (variant && Object.keys(textVariantMapping).includes(variant)) {
     return (
       <Text
         ref={ref}
-        bold={bold}
         color={color as TextProps['color']}
         variant={variant as TextProps['variant']}
+        component={component}
         {...props}
       />
     );
@@ -86,6 +45,7 @@ const Typography = forwardRef(function Typography(
         ref={ref}
         color={color as HeadingProps['color']}
         variant={variant as HeadingProps['variant']}
+        component={component}
         {...props}
       />
     );
@@ -94,14 +54,9 @@ const Typography = forwardRef(function Typography(
 }) as OverridableComponent<TypographyTypeMap>;
 
 export default Typography;
-export { Text };
 export type {
   DefaultTypographyComponent,
   CustomTypographyProps,
   TypographyTypeMap,
   TypographyProps,
-  DefaultTextComponent,
-  CustomTextProps,
-  TextTypeMap,
-  TextProps,
 };
