+++ packages/web-ui/src/Typography/index.ts
@@ -1,12 +1,8 @@
-export { default, textVariantMapping, Text } from './Typography';
+export { default } from './Typography';
 export { typographyThemeOverrides } from './Typography.theme';
 export type {
   DefaultTypographyComponent,
   CustomTypographyProps,
   TypographyTypeMap,
   TypographyProps,
-  DefaultTextComponent,
-  CustomTextProps,
-  TextTypeMap,
-  TextProps,
 } from './Typography';
